LayoutTests/accessibility/alternative-presentation-button-input-type.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../resources/js-test.js"></script>
</head>
<body>
<input type="text" id="textfield">
<p id="description"></p>
<div id="console"></div>
<script>
description("This tests that the alternative presentation button is an accessible element and can be hit test.");

if (window.accessibilityController) {
    debug("<br>Before substitution:");
    var axTextField = accessibilityController.accessibleElementById("textfield");
    dumpRoleAndDescription(axTextField);

    debug("<br>After substitution:");
    window.internals.substituteWithAlternativePresentationButton([document.getElementById("textfield")], 1);
    axTextField = accessibilityController.accessibleElementById("textfield");
    dumpRoleAndDescription(axTextField);

    debug("<br>Hit test alternative presentation button:");
    var x = axTextField.clickPointX;
    var y = axTextField.clickPointY;
    shouldBeTrue("axTextField.elementAtPoint(x, y).isEqual(axTextField)");
}

function dumpRoleAndDescription(accessibleElement)
{
    debug("Role: " + accessibleElement.role);
    debug("Description: " + accessibleElement.description);
}
</script>
</body>
</html>
