LayoutTests/accessibility/alternative-presentation-button-input-type-expected.txt


This tests that the alternative presentation button is an accessible element and can be hit test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Before substitution:
Role: AXRole: AXTextField
Description: AXDescription: 

After substitution:
Role: AXRole: AXButton
Description: AXDescription: 

Hit test alternative presentation button:
PASS axTextField.elementAtPoint(x, y).isEqual(axTextField) is true
PASS successfullyParsed is true

TEST COMPLETE
